+++ src/ConfigProvider.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization-rbac for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization-rbac/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\Expressive\Authorization\Rbac;
 
@@ -19,7 +21,9 @@ class ConfigProvider
     public function getDependencies() : array
     {
         return [
-            ZendRbac::class => ZendRbacFactory::class,
+            'factories' => [
+                ZendRbac::class => ZendRbacFactory::class,
+            ],
         ];
     }
 }