+++ src/ZendRbac.php
@@ -1,14 +1,17 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization-rbac for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization-rbac/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\Expressive\Authorization\Rbac;
 
 use Psr\Http\Message\ServerRequestInterface;
 use Zend\Expressive\Authorization\AuthorizationInterface;
+use Zend\Expressive\Authorization\Exception;
 use Zend\Expressive\Router\RouteResult;
 use Zend\Permissions\Rbac\AssertionInterface;
 use Zend\Permissions\Rbac\Rbac;
@@ -21,7 +24,7 @@ class ZendRbac implements AuthorizationInterface
     private $rbac;
 
     /**
-     * @var AssertionInterface
+     * @var null|AssertionInterface
      */
     private $assertion;