+++ src/ZendRbacFactory.php
@@ -1,14 +1,17 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization-rbac for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization-rbac/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace Zend\Expressive\Authorization\Rbac;
 
 use Psr\Container\ContainerInterface;
 use Zend\Expressive\Authorization\AuthorizationInterface;
+use Zend\Expressive\Authorization\Exception;
 use Zend\Permissions\Rbac\Exception\ExceptionInterface as RbacExceptionInterface;
 use Zend\Permissions\Rbac\Rbac;
 
@@ -43,9 +46,9 @@ class ZendRbacFactory
         $this->injectRoles($rbac, $config['roles']);
         $this->injectPermissions($rbac, $config['permissions']);
 
-        $assertion = $container->has(ZendRbacAssertionInterface::class) ?
-                     $container->get(ZendRbacAssertionInterface::class) :
-                     null;
+        $assertion = $container->has(ZendRbacAssertionInterface::class)
+            ? $container->get(ZendRbacAssertionInterface::class)
+            : null;
 
         return new ZendRbac($rbac, $assertion);
     }